building-image-classifier/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.115.13
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.2.1
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchsummary==1.5.1
torchvision==0.22.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
